Fix tests

+++ tests/test_gpt.py
@@ -57,7 +57,7 @@ def test_gpt_generate(gpt: LM):
     yy = worlds_seqs[:, 1:].contiguous()
     results = gpt(xx, yy=yy, return_logits=True)
     assert results['logits'].shape[1] == max_seq_length - 1
-    log_probs = torch.log_softmax(results['logits'], dim=-1)
+    log_probs = torch.log_softmax(results['logits'], dim=-1).double()
     worlds_log_probs = torch.gather(log_probs, dim=2, index=yy.unsqueeze(dim=2)).squeeze(dim=2)
     worlds_log_probs = torch.sum(worlds_log_probs, dim=1)
     worlds_probs = torch.exp(worlds_log_probs)
